requirements.txt
playwright>=1.48.0
playwright-stealth>=1.0.6
python-slack-sdk>=3.31.0
APScheduler>=3.10.4
pydantic>=2.0.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
